% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{area_chart}
\alias{area_chart}
\title{Create an area chart}
\usage{
area_chart(
  data,
  x,
  y,
  group = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column used for the x-axis, category, or label.}

\item{y}{Column used for the y-axis or value.}

\item{group}{Optional column used to create multiple series.}

\item{...}{Chartkick options passed to JavaScript.}

\item{width, height}{Widget width and height.}

\item{elementId}{Optional htmlwidget element ID.}
}
\description{
Create an area chart
}
