% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chartkick_data}
\alias{chartkick_data}
\title{Convert an R data frame to Chartkick-compatible data}
\usage{
chartkick_data(data, x, y, group = NULL, size = NULL, type = NULL)
}
\description{
Convert an R data frame to Chartkick-compatible data
}
\keyword{internal}
