% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chartkick_timeline_data}
\alias{chartkick_timeline_data}
\title{Convert an R data frame to Chartkick timeline-compatible data}
\usage{
chartkick_timeline_data(data, task, start, end)
}
\description{
Convert an R data frame to Chartkick timeline-compatible data
}
\keyword{internal}
