% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartkickR.R
\name{chartkickR-shiny}
\alias{chartkickR-shiny}
\alias{chartkickROutput}
\alias{renderChartkickR}
\title{Shiny bindings for chartkickR}
\usage{
chartkickROutput(outputId, width = "100\%", height = "400px")

renderChartkickR(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Valid CSS units.}

\item{expr}{Expression that generates a chartkickR widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\description{
Output and render functions for using chartkickR within Shiny applications
and interactive R Markdown documents.
}
