% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{pie_chart}
\alias{pie_chart}
\title{Create a pie chart}
\usage{
pie_chart(data, x, y, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column used for the x-axis, category, or label.}

\item{y}{Column used for the y-axis or value.}

\item{...}{Chartkick options passed to JavaScript.}

\item{width, height}{Widget width and height.}

\item{elementId}{Optional htmlwidget element ID.}
}
\description{
Create a pie chart
}
