% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_data}
\alias{process_data}
\title{Preprocess data frame}
\usage{
process_data(df, group_col, x_col, y_col)
}
\arguments{
\item{df}{A data frame}

\item{group_col}{string value of column name for grouping}

\item{x_col}{string value of column name containing values on the x-axis}

\item{y_col}{string value of column name containing values on the y-axis}
}
\description{
Preprocess data frame
}
\keyword{internal}
