% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{timeline_chart}
\alias{timeline_chart}
\title{Create a timeline chart}
\usage{
timeline_chart(
  data,
  task,
  start,
  end,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{task}{Column containing task or event names.}

\item{start}{Column containing start dates.}

\item{end}{Column containing end dates.}

\item{...}{Chartkick options passed to JavaScript.}

\item{width, height}{Widget width and height.}

\item{elementId}{Optional htmlwidget element ID.}
}
\value{
An htmlwidget object.
}
\description{
Create a timeline chart
}
