% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartkickR.R
\name{chartkickR}
\alias{chartkickR}
\title{Chartkick.js chart in R powered by htmlwidgets}
\usage{
chartkickR(
  data = NULL,
  x,
  y,
  ...,
  type = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data.frame containing data series}

\item{x, y, ...}{List of name value pairs used to map variables on the chart.}

\item{type}{string representing the chart type name i.e. "LineChart"}

\item{width}{chart's width}

\item{height}{chart's height}

\item{elementId}{html element id}
}
\description{
chartkickR is an R package to draw charts based on Chartkick.js JavaScript Library
}
